Update folder and file structure
